Asistant Penjadwalan/backend/requirements.txt
flask==2.3.2
flask-cors==3.0.10
werkzeug==2.3.6
mysql-connector-python==8.0.33
google-generativeai==0.8.3
python-dotenv==1.0.0